initial import of fast open-addressing hashmap
The hashmap uses a tombstone bitmap instead of the special
sentinel keys that eat key space. The bitmap states are:
{ available, occupied, deleted, recycled }

When a node is inserted, it may use an available or delted slot.
In the case of a deleted slot, the node will become recycled.
This is done to keep to simple set and clear bit operations.
When a node is deleted, the occupied bit is cleared.