linkedhashmap: teach insert method to use iterator position hint
previously inserts into the linkedhashmap would always append new
elements to the end of the list, making iteration: insertion order.

change linked list `append_link_internal` to `insert_link_internal`
taking an iterator position, and change the single argument insert
default to inserting before `end()`, which is the same as the
present behaviour, but adds the ability to insert into arbitrary
positions for which one has acquired an iterator.