atomicity change to slot alloction logic in insert
ethical_hashmap does not currently support use of the
containers in multiple threads without locks, however,
it is possible to make insert and delete lock-free in
the case where the container does not need resizing.

this is a subtle change that is not required for
correctness in the single threaded case, however,
the multithreading case has a two-phase 'recycling'
state that is used during deletion to temporarily
mark an entry as both 'occupied' and 'deleted' while
destructors are running. this state means find won't
return the entry because it is marked as deleted
(but may still be occupied) only insert won't reuse
the slot until the occupied bit has been cleared.
this lets us hold a special invariant in the time
window that the destructor is run to keep the entry
in a deleted but not yet avaiable for reuse state.

in insert:

-    if ((state & occupied) != occupied) {
+    if ((state & recycled) == available) {

in find:

     if (state == deleted); /* skip */

in delete:

     bitmap_set(bitmap, i, deleted);
     data[i].~data_type(); /* run destructor */
     bitmap_clear(bitmap, i, occupied);

note: this change is necessary but not sufficient
to make the code thread safe without locks. the
bitmap functions still needs to be updated to use
atomic_compare_exchange and resizng code needs
special consideration.